backend/requirements.txt
Flask>=3.1.2
Flask-SQLAlchemy>=3.1.1
Flask-Cors>=4.0.0
Flask-Migrate>=4.1.0
apscheduler>=3.11.0
python-dateutil>=2.9.0.post0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1